Accept now return StreamConnection
Example server select is now broken.
Accept returns StreamConnection and address is now set in StreamSocket

+++ Noise network/Code/NoiseNetworkTests/Core/DatagramSocketTest.cpp
@@ -68,7 +68,7 @@ namespace NoiseNetworkTests
 			vector<char> buffer;
 			buffer.assign(message.begin(), message.end());
 
-			int result = socket.Send(serverAddress, buffer, buffer.size());
+			int result = socket.Send(serverAddress, buffer, (int)buffer.size());
 			if (result == SOCKET_ERROR)
 			{
 				Assert::Fail(L"Send failed with error code: " + WSAGetLastError());
